mock-interview-backend/requirements.txt
fastapi
uvicorn
langgraph>=1.0.0
langchain>=1.1.0
langchain-core>=1.2.0
langchain-groq>=1.1.0
pydantic>=2.7.4
python-dotenv
